[DU-05] Desenvolvimento dos controllers, utilitários e validadores

+++ desafio-digito-unico-domain/src/main/java/com/desafio/digitounico/dto/DigitoUnicoDTO.java
@@ -9,8 +9,8 @@ public class DigitoUnicoDTO extends BaseDTO {
 
 	private Long id;
 	private String digitoParam;
-	private Long concatenacao;
-	private Long digitoGerado;
+	private Integer concatenacao;
+	private Integer digitoGerado;
 	private Long idUsuario;
 
 }
